refactor(iot): extract checkin workflow into services/iot/checkin.CheckinService (#575 B8)
The 908-line api/iot/checkin/workflow.go held the entire RFID
check-in/checkout state machine in the handler layer. It now lives in
services/iot/checkin: RFID→person→student/staff resolution, checkout and
checkin processing, room/activity-group selection, occupancy counting,
capacity checks, and the WC/Schulhof system-space bootstrap.

Layering: the service returns typed errors (capacity, duplicate-visit
conflict, not-found, internal) and never touches HTTP; handlers keep
device-context validation, request parsing, response building, and the
error→renderer mapping — the PyrePortal-contracted capacity/conflict
wire types stay in api/iot/checkin. Contract strings that moved into the
service are covered by adding services/iot/checkin/checkin_errors.go to
the PyrePortal guard corpus.

External checkin tests keep their assertions unchanged; their
constructor calls follow the new NewResource signature and fault
injection targets the moved resolution seam via a helper. Internal
workflow tests moved to services/iot/checkin. Route/auth-matrix goldens,
wire goldens, PyrePortal guard, and all ratchets green; full suite green.
