+++ backend/api/iot/pyreportal_error_strings_test.go
@@ -131,6 +131,11 @@ var extraGuardSources = []string{
 	"auth/device/errors.go",
 	"services/active/errors.go",
 	"api/common/errors.go",
+	// Issue #575 B8 extracted the RFID check-in business logic into
+	// services/iot/checkin.CheckinService. Its internal-error and
+	// not-found strings (formerly in api/iot/checkin/workflow.go) now live
+	// in this file's message constants; keep it in the guard corpus.
+	"services/iot/checkin/checkin_errors.go",
 }
 
 // TestPyrePortalErrorStringsGuard asserts every PyrePortal-mapped substring
