+++ backend/services/iot/checkin/checkin_system_space.go
@@ -20,7 +20,7 @@ type systemSpace struct {
 	label string // "WC" / "Schulhof" — used in log + error text
 
 	// findRoom returns (nil, nil) when the room does not exist yet.
-	findRoom    func(ctx context.Context, rs *Resource) (*facilities.Room, error)
+	findRoom    func(ctx context.Context, s *CheckinService) (*facilities.Room, error)
 	logRoomName bool // WC logs room_name on found (alias set)
 
 	roomName     string
@@ -36,8 +36,8 @@ type systemSpace struct {
 }
 
 // ensureSystemRoom finds or creates the space's room.
-func (rs *Resource) ensureSystemRoom(ctx context.Context, sp systemSpace) (*facilities.Room, error) {
-	room, err := sp.findRoom(ctx, rs)
+func (s *CheckinService) ensureSystemRoom(ctx context.Context, sp systemSpace) (*facilities.Room, error) {
+	room, err := sp.findRoom(ctx, s)
 	if err != nil {
 		return nil, err
 	}
@@ -46,12 +46,12 @@ func (rs *Resource) ensureSystemRoom(ctx context.Context, sp systemSpace) (*faci
 		if sp.logRoomName {
 			attrs = append(attrs, slog.String("room_name", room.Name))
 		}
-		rs.getLogger().DebugContext(ctx, "found existing "+sp.label+" room", attrs...)
+		s.getLogger().DebugContext(ctx, "found existing "+sp.label+" room", attrs...)
 		return room, nil
 	}
 
 	// Room not found - create it
-	rs.getLogger().InfoContext(ctx, sp.label+" room not found, auto-creating")
+	s.getLogger().InfoContext(ctx, sp.label+" room not found, auto-creating")
 
 	capacity := sp.roomCapacity
 	category := sp.categoryName
@@ -65,30 +65,30 @@ func (rs *Resource) ensureSystemRoom(ctx context.Context, sp systemSpace) (*faci
 		IsSystem: true,
 	}
 
-	if err := rs.FacilityService.CreateRoom(ctx, newRoom); err != nil {
+	if err := s.facilities.CreateRoom(ctx, newRoom); err != nil {
 		// Retry: concurrent request may have created it
-		if room, retryErr := sp.findRoom(ctx, rs); retryErr == nil && room != nil {
+		if room, retryErr := sp.findRoom(ctx, s); retryErr == nil && room != nil {
 			return room, nil
 		}
 		return nil, fmt.Errorf("failed to auto-create %s room: %w", sp.label, err)
 	}
 
-	rs.getLogger().InfoContext(ctx, "successfully auto-created "+sp.label+" room",
+	s.getLogger().InfoContext(ctx, "successfully auto-created "+sp.label+" room",
 		slog.Int64("room_id", newRoom.ID),
 	)
 	return newRoom, nil
 }
 
 // ensureSystemCategory finds or creates the space's activity category.
-func (rs *Resource) ensureSystemCategory(ctx context.Context, sp systemSpace) (*activities.Category, error) {
-	categories, err := rs.ActivitiesService.ListCategories(ctx)
+func (s *CheckinService) ensureSystemCategory(ctx context.Context, sp systemSpace) (*activities.Category, error) {
+	categories, err := s.activities.ListCategories(ctx)
 	if err != nil {
 		return nil, fmt.Errorf("failed to list activity categories: %w", err)
 	}
 
 	for _, cat := range categories {
 		if cat.Name == sp.categoryName {
-			rs.getLogger().DebugContext(ctx, "found existing "+sp.label+" category",
+			s.getLogger().DebugContext(ctx, "found existing "+sp.label+" category",
 				slog.Int64("category_id", cat.ID),
 			)
 			return cat, nil
@@ -96,7 +96,7 @@ func (rs *Resource) ensureSystemCategory(ctx context.Context, sp systemSpace) (*
 	}
 
 	// Category not found - create it
-	rs.getLogger().InfoContext(ctx, sp.label+" category not found, auto-creating")
+	s.getLogger().InfoContext(ctx, sp.label+" category not found, auto-creating")
 
 	newCategory := &activities.Category{
 		Name:        sp.categoryName,
@@ -105,10 +105,10 @@ func (rs *Resource) ensureSystemCategory(ctx context.Context, sp systemSpace) (*
 		IsSystem:    true,
 	}
 
-	createdCategory, err := rs.ActivitiesService.CreateCategory(ctx, newCategory)
+	createdCategory, err := s.activities.CreateCategory(ctx, newCategory)
 	if err != nil {
 		// Retry: concurrent request may have created it
-		retryCategories, retryErr := rs.ActivitiesService.ListCategories(ctx)
+		retryCategories, retryErr := s.activities.ListCategories(ctx)
 		if retryErr == nil {
 			for _, cat := range retryCategories {
 				if cat.Name == sp.categoryName {
@@ -119,7 +119,7 @@ func (rs *Resource) ensureSystemCategory(ctx context.Context, sp systemSpace) (*
 		return nil, fmt.Errorf("failed to auto-create %s category: %w", sp.label, err)
 	}
 
-	rs.getLogger().InfoContext(ctx, "successfully auto-created "+sp.label+" category",
+	s.getLogger().InfoContext(ctx, "successfully auto-created "+sp.label+" category",
 		slog.Int64("category_id", createdCategory.ID),
 	)
 	return createdCategory, nil
@@ -127,13 +127,13 @@ func (rs *Resource) ensureSystemCategory(ctx context.Context, sp systemSpace) (*
 
 // systemActivityGroup finds or creates the space's permanent activity group,
 // lazily bootstrapping room + category + activity on first use.
-func (rs *Resource) systemActivityGroup(ctx context.Context, sp systemSpace) (*activities.Group, error) {
+func (s *CheckinService) systemActivityGroup(ctx context.Context, sp systemSpace) (*activities.Group, error) {
 	var room *facilities.Room
 	var err error
 	if sp.selectActivity != nil {
 		// Selectors such as Schulhof need the canonical room to distinguish the
 		// dedicated system activity from normal activities with the same name.
-		room, err = rs.ensureSystemRoom(ctx, sp)
+		room, err = s.ensureSystemRoom(ctx, sp)
 		if err != nil {
 			return nil, fmt.Errorf("failed to ensure %s room: %w", sp.label, err)
 		}
@@ -146,7 +146,7 @@ func (rs *Resource) systemActivityGroup(ctx context.Context, sp systemSpace) (*a
 	filter.Equal("name", sp.activityName)
 	options.Filter = filter
 
-	groups, err := rs.ActivitiesService.ListGroups(ctx, options)
+	groups, err := s.activities.ListGroups(ctx, options)
 	if err != nil {
 		return nil, fmt.Errorf("failed to query %s activity: %w", sp.label, err)
 	}
@@ -162,23 +162,23 @@ func (rs *Resource) systemActivityGroup(ctx context.Context, sp systemSpace) (*a
 	}
 
 	if existingActivity := selectExisting(groups); existingActivity != nil {
-		rs.getLogger().DebugContext(ctx, "found existing "+sp.label+" activity",
+		s.getLogger().DebugContext(ctx, "found existing "+sp.label+" activity",
 			slog.Int64("activity_id", existingActivity.ID),
 		)
 		return existingActivity, nil
 	}
 
 	// Activity not found - auto-create the entire infrastructure
-	rs.getLogger().InfoContext(ctx, sp.label+" activity not found, auto-creating infrastructure")
+	s.getLogger().InfoContext(ctx, sp.label+" activity not found, auto-creating infrastructure")
 
 	if room == nil {
-		room, err = rs.ensureSystemRoom(ctx, sp)
+		room, err = s.ensureSystemRoom(ctx, sp)
 		if err != nil {
 			return nil, fmt.Errorf("failed to ensure %s room: %w", sp.label, err)
 		}
 	}
 
-	category, err := rs.ensureSystemCategory(ctx, sp)
+	category, err := s.ensureSystemCategory(ctx, sp)
 	if err != nil {
 		return nil, fmt.Errorf("failed to ensure %s category: %w", sp.label, err)
 	}
@@ -194,14 +194,14 @@ func (rs *Resource) systemActivityGroup(ctx context.Context, sp systemSpace) (*a
 	}
 
 	// CreateGroup requires supervisorIDs and schedules - pass empty slices for auto-created activity
-	createdActivity, err := rs.ActivitiesService.CreateGroup(ctx, newActivity, []int64{}, []*activities.Schedule{})
+	createdActivity, err := s.activities.CreateGroup(ctx, newActivity, []int64{}, []*activities.Schedule{})
 	if err != nil {
 		// Retry: concurrent request may have created it
 		retryOptions := base.NewQueryOptions()
 		retryFilter := base.NewFilter()
 		retryFilter.Equal("name", sp.activityName)
 		retryOptions.Filter = retryFilter
-		retryGroups, retryErr := rs.ActivitiesService.ListGroups(ctx, retryOptions)
+		retryGroups, retryErr := s.activities.ListGroups(ctx, retryOptions)
 		if retryErr == nil {
 			if existingActivity := selectExisting(retryGroups); existingActivity != nil {
 				return existingActivity, nil
@@ -210,7 +210,7 @@ func (rs *Resource) systemActivityGroup(ctx context.Context, sp systemSpace) (*a
 		return nil, fmt.Errorf("failed to auto-create %s activity: %w", sp.label, err)
 	}
 
-	rs.getLogger().InfoContext(ctx, "successfully auto-created "+sp.label+" infrastructure",
+	s.getLogger().InfoContext(ctx, "successfully auto-created "+sp.label+" infrastructure",
 		slog.Int64("room_id", room.ID),
 		slog.Int64("category_id", category.ID),
 		slog.Int64("activity_id", createdActivity.ID),
